Rename scenarios/aws/vm/os to ec2os
