+++ scenarios/aws/vm/ec2os/fedora.go
@@ -1,4 +1,4 @@
-package os
+package ec2os
 
 import (
 	"errors"
