+++ scenarios/aws/vm/ec2os/os.go
@@ -1,4 +1,4 @@
-package os
+package ec2os
 
 import (
 	"fmt"
