+++ scenarios/aws/vm/ec2os/windows.go
@@ -1,4 +1,4 @@
-package os
+package ec2os
 
 import (
 	"errors"
